+++ src/utils.js
@@ -6,7 +6,7 @@ var isLocation = !!(window.history.location || window.location);
  * @param {boolean} hash Set to true when using hash routing
  * @returns {bool} true if the link is valid for navigation, false otherwise
  */
-function validLink(el, hash) {
+export function validLink(el, hash) {
 	// The checks in this block are taken from page.js https://github.com/visionmedia/page.js/blob/master/index.js#L370
 	// el.nodeName for svg links are 'a' instead of 'A'
 	while (el && 'A' !== el.nodeName.toUpperCase()) el = el.parentNode;
@@ -70,98 +70,12 @@ export function processTrailingSlash(path, trail) {
 }
 
 /**
- * @param {boolean} hash set to  true when using hash routing
- * @param {object} render render settings
- * @description Add a handler to click events on all links currently in the page
- * if using page or views rendering this will be called everytime the page changes
- * this may also be called by the developer if they added other links dynamicly
+ * Create the context object
+ * @param {string} route the route
+ * @param {string} path the path accessed by the client
+ * @param {array} props the routes variables
+ * @returns {object} the context object
  */
-export function interceptLinks(hash, render = null) {
-	document.querySelectorAll('a').forEach((el) => {
-		// check if we already add this link
-		if (el.hasAttribute('x-link')) return;
-		// check if the link is a navigation link
-		if (validLink(el, hash) == false) return;
-
-		// add an x-link attribute this will tell this function
-		// that the link already been handled.
-		el.setAttribute('x-link', '');
-
-		// X-RENDER ONLY
-		if (render != null) {
-			el.addEventListener('mouseover', (e) => {
-				if (!render.enabled || !render.preload) {
-					return;
-				}
-				let path = e.target.getAttribute('href');
-				if (path == null) path = '/';
-				if (render.preloaded.path == path) {
-					return;
-				}
-
-				window.setTimeout(function () {
-					fetch(path)
-						.then((response) => {
-							return response.text();
-						})
-						.then((response) => {
-							render.preloaded.path = path;
-							render.preloaded.content = response;
-						});
-				});
-			});
-		}
-		// X-RENDER END
-
-		el.addEventListener(
-			'click',
-			(e) => {
-				e.preventDefault();
-				let link = el.pathname;
-				if (hash) {
-					window.location.hash = '#' + link;
-				} else {
-					AlpineRouter.navigate(link);
-				}
-			},
-			false
-		);
-	});
-}
-
-/**
- * This will replace the content fetched from `path` into `selector`.
- * The content is assumed to not be an entire html page but a chunk of it.
- * @param {string} content the html content.
- * @param {string} selector the selector of where to put the content.
- */
-export function renderContent(content, selector) {
-	// replace the content of the selector with the fetched content
-	document.querySelector(selector).innerHTML = content;
-}
-
-/**
- * This will replace the content fetched from `path` into `selector`.
- * Unlike renderContent, this will assume the fetched content to be an entire HTML.
- * meaning it needs to process the routes as well.
- * @param {string} content the html content.
- * @param {string} selector the selector of where to put the content.
- * @param {array} routes routes array to be processed.
- * @returns {array} processed routes
- */
-export function renderPage(content, selector, routes) {
-	let doc = new DOMParser().parseFromString(content, 'text/html');
-	doc = doc.querySelector(selector);
-	// This takes the document fetched, remove routers already initialized from it
-	// and also remove routers initialized but not found in it
-	// that is for routers that are not needed in this page.
-	let r = processRoutersInFetchedDoc(doc, selector, routes);
-	doc = r.doc;
-	content = doc.innerHTML;
-	renderContent(content, selector);
-	return r.routes;
-}
-
 export function buildContext(route, path, props) {
 	return {
 		route: route,
@@ -169,69 +83,63 @@ export function buildContext(route, path, props) {
 		props: props,
 		query: window.location.search.substring(1), // query w/out leading '?'
 		hash: window.location.hash.substring(1), // hash without leading '#'
-		go: (path) => {
-			window.AlpineRouter.navigate(path);
+		redirect: (path) => {
+			window.PineconeRouter.navigate(path);
 			return false; // returning false will stop the navigate function, before rendering any views etc.
-		},
-		setHash(hash) {
-			window.location.hash = hash;
-		},
-		setQuery(q) {
-			window.location.search = q;
 		},
 	};
 }
 
 /**
- * This takes the document fetched, remove routers already initialized from it
- * @param {Document} doc
- * @param {array} routes
- * @returns {object} {doc, routes}
+ * Call a function on all middlewares loaded, if any.
+ * @param {string} func middlware function to call.
+ * @param {array} args arguments to pass to the function.
+ * @returns {boolean} false if the middleware function return false, i.e. it want to stop execution of the function and return.
  */
-function processRoutersInFetchedDoc(doc, selector, routes) {
-	let routersInDoc = doc.querySelectorAll('[x-router]');
-	switch (routersInDoc.length) {
-		case 0:
-			// if there is no router in the fetched doc, remove the routes registered
-			// but only if the selector is body
-			if (selector == 'body') routes = [];
-			break;
-		case 1:
-			// the router currently loaded
-			let currentRouter = document.querySelector('[x-router]');
-			// if the router in the doc dont have x-router set to 'loaded'
-			// thus remove it from the current router element before checking if they're the same
-			currentRouter.setAttribute('x-router', '');
-			// check if the one in fetched doc is the same as the current one
-			if (
-				routersInDoc[0].isEqualNode(
-					document.querySelector('[x-router]')
-				)
-			) {
-				// if it is, mark the router as loaded, so routes wont be processed again
-				routersInDoc[0].setAttribute('x-router', 'loaded');
-				// remove the router element currently in the page, in case it is not within the selector.
-				document.querySelector('[x-router]').remove();
-			} else {
-				// if they're not the same remove the routes, the new ones will be added once this new router is added
-				routes = [];
-				document.querySelector('[x-router]').remove();
+export function middleware(func, args) {
+	if (window.PineconeRouterMiddlewares == null) return;
+	window.PineconeRouterMiddlewares.forEach((plugin) => {
+		if (plugin[func] == null) return;
+		return plugin[func](...args);
+	});
+}
+
+/**
+ * execute the handlers of routes that are given passing them the context.
+ * @param {array} handlers handlers to execute.
+ * @param {array} context the current context to pass as argument.
+ * @returns {boolean} false if the handler request a redirect.
+ */
+export function handle(handlers, context) {
+	switch (typeof handlers) {
+		// multiple handlers; typeof array == 'object
+		case 'object':
+			if (handlers.length == 0) return;
+
+			for (const i in handlers) {
+				if (typeof handlers[i] == 'function') {
+					let result = handlers[i](context);
+					// if the handler redirected, return
+					// if not go to the next handler
+					if (result == false) {
+						return false;
+					}
+				}
 			}
 			break;
-		default:
-			// more than one
-			throw new Error(
-				'Alpine Router: there can only be one router in the same page'
-			);
-	}
 
-	return { doc, routes };
+		// single handler
+		case 'function':
+			return handlers(context);
+	}
 }
 
 /**
  * Match the path with specified routes
  * Taken from https://github.com/vijitail/simple-javascript-router/blob/master/src/router/Router.js#L14
  * with some modifications to fix bad behavior
+ * @param {string} route the route to check for.
+ * @param {string} requestPath the path that's requested by the client.
  */
 export function match(route, requestPath) {
 	let paramNames = [];
@@ -242,19 +150,18 @@ export function match(route, requestPath) {
 			return '([^/]+)';
 		}) + '(?:/|$)';
 
-	let params = {};
 	let routeMatch = requestPath.match(new RegExp(regexPath));
 	if (routeMatch !== null) {
 		if (routeMatch.index != 0) return null;
 		if (routeMatch.input != routeMatch[0]) return null;
-		params = routeMatch.slice(1).reduce((params, value, index) => {
+		let props = routeMatch.slice(1).reduce((params, value, index) => {
 			if (params === null) params = {};
 			params[paramNames[index]] = value;
 			return params;
 		}, null);
+		route.setProps(props);
 	}
 
-	route.setProps(params);
 	return routeMatch;
 }